+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- ME2LE: Added Legendary missing data (ME1 Import Rewards)
+
 ## 1.11.0
 - Added a command line tool to compare the plots of 2 saves, additional information with the --help argument
 - Added a `Reload` button for those who often need to reload there saves (modders, etc.)